feat(team-run-flow): per-worker tool calls in the run-flow drawer
Reuses the single-agent tool-call view for team mode:

- gateway: the team worker loop's onStatus was a no-op, dropping every
  worker's tool events. Forward tool_start/tool_end to the chat sink
  (mirrors the single-agent onStatus). Parallel path left untouched —
  its concurrent tool events can't be attributed by order.
- Mac: extract the tool-timeline renderer into a shared <ToolCallList>
  (ChatContextPanel's ToolTimeline now wraps it, so single-agent is
  literally the same component). teamRunModel.toolCallsForStep() buckets
  each tool call under the most-recent `Step N` marker (team runs are
  serial), and TeamRunFlow's drawer renders that worker's calls.
